requirements.txt
feast_custom_provider/
pytest==6.2.4
pyspark
koalas